Create :chpl:chplref:`chplmodindex` role to cross-reference chapel module index.
It would be ideal to support something like :ref:`chplmodindex`, but it is
difficult to expand "ref" without re-implementing it. Since this is a one-off
cross-reference to a meta-page, it seems OK for it to be a little more
cumbersome and less intuitive. Especially since it means :ref:`modindex`,
:ref:`genindex`, etc will continue to work as expected without reimplementing
support for them.