Update signature patterns to better match return types (and drop -> before return type).
Instead of trying to pick out the type that occurs after the colon, or after a
space if its a ref return intent, pick everything including the colon or
space. This will then be used in the output as the return annotation.

E.g., if the signature is `x():int`, then the return annotation will literally
be `:int`. Similarly:

`x(int(64)) : int(32)` --> ` : int(32)`
`x()   const`          --> `   const`

Also, update the node type for the return annotation to simply be
`addnodes.desc_type` instead of `addnodes.desc_returns`. desc_returns is a
subclass of desc_type that overrides the astext() method and inserts the string
" -> " before the actual text of the node.

By using desc_type directly, there will no longer be the " -> " string. But,
now that the return annotations preserve the colon, those will be in the node
text and show up in the output (i.e. html).