Fix output for type and param methods
Before, the Sphinx output for methods of the form `proc type foo` and `proc
param foo` (the latter of which was only applicable to types which support
param, such as primitives like ints) would get somewhat messed up, as the domain
wasn't aware that these were prefix keywords and tried to treat them like names
for the methods.  This also broke linking to the methods.  With this change,
the Sphinx output should be improved.

Also added test coverage for the new cases.