Add support for a :throw(s): tag for functions and iterators
The formatting is similar to :arg name: (because that's how the example I saw
did it for Python errors), but it doesn't allow a :type name: extension
(because that seemed silly to me).  Throw lists should be after both arguments
and return types (because that's how it is in Chapel function declarations).

Example - for a function defined as:

```chapel
proc foo(x: int): string throws { ... }
```

We expect to write the following documentation:
```
Something about foo

:arg x: some number
:type x: `int`
:return: some message
:rtype: `string`
:throws SomeError: should some condition occur
:throws SomeDifferentError: should some other condition occur
```