Update Python versions used by Travis
Stop testing Python 2 at all, we want to drop support for it.  Also, update the
version of Python 3 being used, as the latest Sphinx only supports Python 3.5
and up (I am going with the lowest one supported for now, but it seems
reasonable to require a later one if we so choose).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>