Fix release script for new standard.
Using setup.py's upload command is not the preferred way to upload to pypi now,
we need to use twine.  So update the script to install twine and wheel, and use
the commands twine expects in order to upload to pypi.

There may also be an issue with uploading via the script for authentication
purposes?  Either that or I typed my password wrong when I ran it through the
script.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>