Adjust __init__.py for Sphinx's new requirements
make_field is expected to have a different signature, so I updated it to reflect
the one in sphinx/util/docfields.py (the function was a copy of a previous
version, so this seemed like the most appropriate course of action).  This
necessitated some import updates.  This change gets us further to passing
tox testing, but there are now some failures because we are passing a "None"
environment to make_xrefs and it doesn't expect that.

Still investigating, but snapshotting the changes.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>