Migrate to github actions
Migrate from Travis to Github Actions. Travis is no longer free (or well
supported), so migrate to github actions. For the most part this is a
direct transliteration of the Travis job, but the main coveralls
github-action doesn't work with xml (coverallsapp/github-action 30), so
switch to using codecov, which provides similar functionality.

Here's some resources I used:
- https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
- https://github.com/codecov/codecov-action

Resolves Cray/chapel-private 1804