Switch from nosetest to pytest
nosetest hasn't been updated/supported in years and no longer works with
python 3.9+ (nose-devs/nose 1099). Switch to pytest and upgrade python
versions.

Note that this appears to drop our code coverage from ~60% to ~40%, but
that's because nosetest was including coverage results for the tests
themselves, which isn't particularly helpful anyways.

As part of this, switch CI to call tox to run testing instead of copying
the commands we use.