Update setuptools used versions and set pytest to more recent version
Looks like the problem is coming from more recent versions of packages that we
pin for our requirements files.  setuptools doesn't use the requirements files
to install dependencies, so it needs the limits as well.

While there, since the problem didn't seem to be pytest, pin pytest to the more
recent version

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>