Update the version and add a note to the requirements file
The version update is the usual post-release version bump so that we are ready
for the next one.

The requirements file note is so that we don't forget to update setup.py the
next time we update the requirements.txt file.  We had previously removed the
version numbers from setup.py under the mistaken belief that setting it in the
requirements file was sufficient to affect what was used when calling setup.py.
However, that was not the case in the independent release job, so it needed the
version numbers restored

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>